Update sidebar styling and transition
