+++ src/App.css
@@ -22,25 +22,30 @@ h6 {
   font-family: var(--heading-font);
 }
 
+.column,
 .row {
   display: flex;
 }
 
 .column {
-  display: flex;
   flex-direction: column;
 }
 
 section {
   padding: 4rem 0;
   min-height: 100vh;
-  width: 80vw;
+  width: 100%;
 }
 
 .background {
-  background: url(/public/images/florence-italy.jpg), var(--theme-lime);
+  background: url(/public/images/pattern.jpg), var(--theme-lime);
+  background: url(/public/images/balloons.jpg), var(--theme-lime);
+  background: url(/public/images/green.jpg), var(--theme-lime);
   background-repeat: no-repeat;
   background-size: cover;
+  background-blend-mode: luminosity;
+  background-blend-mode: soft-light;
+  background-blend-mode: soft-light;
   background-blend-mode: multiply;
 }
 
@@ -100,7 +105,7 @@ section {
 
 /* Sidebar */
 .side-bar-container {
-  align-items: center;
+  /* align-items: center; */
   justify-content: space-between;
   border-right: #8c8c8c 1px solid;
   width: 20vw;
@@ -144,9 +149,10 @@ section {
 
 /* Carausel / Transitions */
 .events-list {
-  --diameter: 200px;
-  --growth: 50px;
+  --diameter: calc(calc(calc(80vw - 172px) / 3) - 4rem);
+  --growth: 4rem;
 
+  position: relative;
   align-items: flex-start;
   justify-content: space-between;
   width: 100%;
@@ -154,38 +160,64 @@ section {
 
 .card-enter {
   opacity: 0;
-  width: 0;
   height: 0;
   margin: 0;
   padding: 0;
 }
 
+.card-enter .circle {
+  width: 0;
+  border-width: 1px;
+}
+
 .card-enter-active {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(50%, 0%);
   opacity: 1;
-  width: 232px;
-  height: 292px;
+  height: calc(var(--diameter) + 2rem + 60px);
   margin: 1rem;
   padding: 1rem;
-  transition: height 2s ease-in-out, width 5s ease-in-out, opacity 3s ease-in-out;
+  transition: height 2s ease-in-out, opacity 4s ease-in-out;
+}
+
+.card-enter-active .circle {
+  width: var(--diameter);
+  border-width: 4px;
+  transition: all 2s ease-in-out;
+}
+
+.card-enter-active:hover {
+  pointer-events: none !important;
 }
 
 .card-exit {
   opacity: 1;
-  width: 232px;
-  height: 292px;
+  height: calc(var(--diameter) + 1rem + 60px);
   margin: 1rem;
   padding: 1rem;
 }
 
+.card-exit .circle {
+  width: var(--diameter);
+  border-width: 4px;
+}
+
 .card-exit-active {
+  z-index: -1;
   opacity: 0;
-  width: 0;
   height: 0;
   margin: -2rem !important;
   padding: 0 !important;
   border-radius: 50%;
-  overflow: hidden;
-  transition: height 5s ease-in-out, width 1s ease-in-out, opacity 5s ease-in-out;
+  transition: all 2s ease-in-out, opacity 1s ease-in-out;
+}
+
+.card-exit-active .circle {
+  width: 0;
+  border-width: 1px;
+  transition: all 2s ease-in-out;
 }
 
 /* Event */
@@ -210,6 +242,7 @@ section {
 .event-card:hover .circle {
   height: calc(var(--diameter) + var(--growth));
   width: calc(var(--diameter) + var(--growth));
+  margin: -1rem;
   border-radius: 1rem;
   box-shadow: 0 0 0 8px white inset;
 }
@@ -358,25 +391,25 @@ hr {
 /* Add event page */
 
 .add-event-main {
-  display: flex;
-  flex-direction: column;
   align-items: center;
-  justify-content: flex-start;
-  width: 100%;
-  height: 100vh;
-  background: url(/public/images/musical_festival.jpg), var(--theme-lime);
-  background-repeat: no-repeat;
-  background-size: cover;
+
+  /* justify-content: flex-start; */
+
+  /* width: 100%; */
+
+  /* height: 100%; */
+
+  /* background: url(/public/images/musical_festival.jpg), var(--theme-lime); */
+
+  /* background-repeat: no-repeat; */
+
+  /* background-size: cover; */
   background-blend-mode: soft-light;
 }
 
 .add-event-form {
-  display: flex;
-  flex-direction: column;
   align-items: center;
-  justify-content: flex-start;
-  width: 100%;
-  height: 100vh;
+  min-height: 100vh;
 }
 
 #submit-event {
@@ -405,18 +438,14 @@ input::file-selector-button {
 /* About page */
 
 .authors-div-body {
-  display: flex;
   align-items: center;
   justify-content: space-around;
 }
 
 .about-main-container {
-  display: flex;
-  flex-direction: column;
   align-items: center;
   justify-content: center;
-  width: 80%;
-  height: 100vh;
+  width: 100%;
 }
 
 .about-text {
@@ -424,6 +453,14 @@ input::file-selector-button {
   font-weight: 500;
   text-align: center;
   margin: 1rem 3rem;
+}
+
+.authors-div-body-item {
+  width: 30%;
+}
+
+.authors-div-body-item img {
+  width: 90%;
 }
 
 .about-icons {
